+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Revision history for squares
 
+## 0.0.1 -- 2020-05-23
+
+* Included README.
+* Added `Control.Arrow.Square`.
+* Removed use of BlockArguments to support older GHC versions.
+
 ## 0 -- 2020-05-22
 
 * First version. Released on an unsuspecting world.